% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dpCluster.R
\name{dpCluster}
\alias{dpCluster}
\title{dpCluster}
\usage{
dpCluster(x, percent = 0.01, thres.rho = NULL, thres.delta = NULL,
  similarity = c("euclidean", "SNN"), method = c("gaussian", "withinDc",
  "neighbors"), threads = 4, halo.detection = TRUE)
}
\arguments{
\item{x}{data.frame or matrix}

\item{percent}{method `gaussian` and `withinDc` use the nearest `percent` of points to estimate `dc`;

       method `neighbors` use the nearest `percent` of points to estimate density directly.

       percent is suggested between 0.01 and 0.02}

\item{thres.rho}{threshold of rho to detect peaks, which is often get from "Decision Graph".}

\item{thres.delta}{threshold of delta to detect peaks, which is often get from "Decision Graph".}

\item{method}{there are three methods to estimate density.

       `gaussian` : gaussian kernel density estimation $e^{(\frac{-1}{2}(\frac{\delta(x_{i}, x_{j})}{2dc})^{2})}$

       `withinDc` : the density of $x_{i}$ is defined as the number of points within dc distance of $x_{i}$

       `neighbors` : the density of $x_{i}$ is defined as the reciprocal of the mean distance of $x_{i}$'s neighbors}

\item{threads}{number of threads to use for task scheduling}

\item{halo.detection}{logical, whether to remove the potential noise point with low density}
}
\description{
dpCluster
}

